@@ -1,4 +1,5 @@
 import * as React from "react";
+import {FormattedMessage} from "react-intl";
 import {style} from "typestyle";
 import {Link} from "./Link";
 
@@ -23,10 +24,11 @@ const Styles = {
 const Header = () => (
   <nav className={Styles.nav}>
     <ul>
-      <li><Link name="home">Home</Link></li>
-      <li><Link name="about">About</Link></li>
-      <li><Link name="counter">Counter</Link></li>
-      <li><Link name="stars">Stars</Link></li>
+      <li><Link name="home"><FormattedMessage id="home" defaultMessage="Home" /></Link></li>
+      <li><Link name="about"><FormattedMessage id="about" defaultMessage="About Us" /></Link></li>
+      <li><Link name="counter"><FormattedMessage id="counter" defaultMessage="Counter" /></Link></li>
+      <li><Link name="stars"><FormattedMessage id="stars" defaultMessage="Stars" /></Link></li>
+      <li><Link name="register"><FormattedMessage id="register" defaultMessage="Register" /></Link></li>
     </ul>
   </nav>
 );
